ensure pstutil for the process is accurate
